27/3 9:43 Comments added. Print-outs removed. Factory method moved. Final version.
