+++ assignment2_chat/chat_server/server/src/mediator/CommunicationPackage.java
@@ -1,4 +1,4 @@
-package model;
+package mediator;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
