+++ assignment2_chat/chat_server/server/src/mediator/Package.java
@@ -1,4 +1,4 @@
-package model;
+package mediator;
 
 public abstract class Package {
