+++ assignment2_chat/chat_server/server/src/mediator/UserPackage.java
@@ -1,4 +1,4 @@
-package model;
+package mediator;
 
 public class UserPackage extends Package
 {
